prevent auto-run of app on test
